+++ source_gen/CHANGELOG.md
@@ -4,6 +4,10 @@
   do resolution need to switch to the new API, see
   https://github.com/dart-lang/sdk/blob/main/pkg/analyzer/doc/element_model_migration_guide.md.
   For questions please use https://github.com/dart-lang/build/discussions.
+- **Breaking Change**: annotations are now reported separately for elements
+  and directives. If you use `TypeChecker().annotatedWith()`, see also
+  the new `TypeChecker().libraryDirectivesAnnotatedWith()`. If you use
+  `GeneratorForAnnotation` see the new `generateForAnnotatedDirective`.
 - Require `build: ^3.0.0`
 - Updated the minimum package versions for a number of dependencies.
 - Require Dart 3.7.0